requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlmodel==0.0.14
pydantic==1.10.13
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
fastapi-mail==1.4.1
APScheduler==3.10.4
tzlocal==5.2